+++ src/countries/iq.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const iq: Country = {
+	code: 'iq',
+	name: {
+		en: 'Iraq',
+		ar: 'العراق',
+		de: 'Irak',
+		sv: 'Irak',
+		da: 'Irak',
+		sq: 'Iraku',
+		pt: 'Iraque',
+		tr: 'Irak',
+		fa: 'عراق',
+		uk: 'Ірак',
+		xa: 'موطني',
+		xt: 'موطني',
+	},
+	flag: '🇮🇶',
+	beta: true,
+}