+++ src/countries/lb.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const lb: Country = {
+	code: 'lb',
+	name: {
+		en: 'Lebanon',
+		ar: 'لبنان',
+		de: 'Libanon',
+		sv: 'Libanon',
+		da: 'Libanon',
+		sq: 'Libani',
+		pt: 'Líbano',
+		tr: 'Lübnan',
+		fa: 'لبنان',
+		uk: 'Ліван',
+		xa: 'كلّنا للوطن',
+		xt: 'كلّنا للوطن',
+	},
+	flag: '🇱🇧',
+	beta: true,
+}